+++ src/py42/_internal/client_factories.py
@@ -2,13 +2,12 @@ import json
 
 from requests import HTTPError
 
-from py42._internal.clients import archive
-from py42._internal.clients import key_value_store
-from py42._internal.clients import securitydata
-from py42.clients import administration, alerts, devices, legalhold, orgs, users
+from py42._internal.clients import alerts, archive, key_value_store, securitydata
+from py42._internal.clients.detection_list_user import DetectionListUserClient
+from py42.clients import administration, devices, legalhold, orgs, users
+from py42._internal.clients.alertrules import AlertRulesClient
 from py42.clients.detectionlists.departing_employee import DepartingEmployeeClient
 from py42.clients.detectionlists.high_risk_employee import HighRiskEmployeeClient
-from py42._internal.clients.detection_list_user import DetectionListUserClient
 from py42.clients.file_event import FileEventClient
 from py42.exceptions import Py42FeatureUnavailableError, Py42SessionInitializationError
 
@@ -46,19 +45,23 @@ class MicroserviceClientFactory(object):
         root_session,
         session_factory,
         user_context,
+        user_client,
         key_value_store_client=None,
     ):
         self._authority_url = authority_url
         self._root_session = root_session
         self._session_factory = session_factory
         self._user_context = user_context
+        self._user_client = user_client
         self._key_value_store_client = key_value_store_client
+
         self._alerts_client = None
         self._departing_employee_client = None
         self._file_event_client = None
         self._high_risk_employee_client = None
         self._detection_list_user_client = None
         self._ecm_session = None
+        self._alert_rules_client = None
 
     def get_alerts_client(self):
         if not self._alerts_client:
@@ -66,14 +69,12 @@ class MicroserviceClientFactory(object):
             self._alerts_client = alerts.AlertClient(session, self._user_context)
         return self._alerts_client
 
-    def get_departing_employee_client(self, user_client):
+    def get_departing_employee_client(self):
         if not self._departing_employee_client:
             if not self._ecm_session:
                 self._ecm_session = self._get_jwt_session(u"employeecasemanagement-API_URL")
             self._departing_employee_client = DepartingEmployeeClient(
-                self._ecm_session,
-                self._user_context,
-                self.get_detection_list_user_client(user_client),
+                self._ecm_session, self._user_context, self.get_detection_list_user_client()
             )
         return self._departing_employee_client
 
@@ -82,6 +83,33 @@ class MicroserviceClientFactory(object):
             session = self._get_jwt_session(u"FORENSIC_SEARCH-API_URL")
             self._file_event_client = FileEventClient(session)
         return self._file_event_client
+
+    def get_high_risk_employee_client(self):
+        if not self._high_risk_employee_client:
+            if not self._ecm_session:
+                self._ecm_session = self._get_jwt_session(u"employeecasemanagement-API_URL")
+            self._high_risk_employee_client = HighRiskEmployeeClient(
+                self._ecm_session, self._user_context, self.get_detection_list_user_client()
+            )
+        return self._high_risk_employee_client
+
+    def get_detection_list_user_client(self):
+        if not self._detection_list_user_client:
+            if not self._ecm_session:
+                self._ecm_session = self._get_jwt_session(u"employeecasemanagement-API_URL")
+            user_client = self._user_client
+            self._detection_list_user_client = DetectionListUserClient(
+                self._ecm_session, self._user_context, user_client
+            )
+        return self._detection_list_user_client
+
+    def get_alert_rules_client(self):
+        if not self._alert_rules_client:
+            session = self._get_jwt_session(u"FedObserver-API_URL")
+            self._alert_rules_client = AlertRulesClient(
+                session, self._user_context, self.get_detection_list_user_client()
+            )
+        return self._alert_rules_client
 
     def _get_jwt_session(self, key):
         url = self._get_stored_value(key)
@@ -93,26 +121,6 @@ class MicroserviceClientFactory(object):
             session = self._session_factory.create_anonymous_session(url)
             self._key_value_store_client = key_value_store.KeyValueStoreClient(session)
         return self._key_value_store_client.get_stored_value(key).text
-
-    def get_high_risk_employee_client(self, user_client):
-        if not self._high_risk_employee_client:
-            if not self._ecm_session:
-                self._ecm_session = self._get_jwt_session(u"employeecasemanagement-API_URL")
-            self._high_risk_employee_client = HighRiskEmployeeClient(
-                self._ecm_session,
-                self._user_context,
-                self.get_detection_list_user_client(user_client),
-            )
-        return self._high_risk_employee_client
-
-    def get_detection_list_user_client(self, user_client):
-        if not self._detection_list_user_client:
-            if not self._ecm_session:
-                self._ecm_session = self._get_jwt_session(u"employeecasemanagement-API_URL")
-            self._detection_list_user_client = DetectionListUserClient(
-                self._ecm_session, self._user_context, user_client
-            )
-        return self._detection_list_user_client
 
 
 def _hacky_get_microservice_url(session, microservice_base_name):