+++ tests/_internal/test_archive_access.py
@@ -17,8 +17,7 @@ from py42._internal.archive_access import (
     RestoreJobManager,
 )
 from py42._internal.clients.archive import ArchiveClient
-from py42._internal.clients.storage import StorageArchiveClient
-from py42._internal.clients.storage import StorageClient, StorageClientFactory
+from py42._internal.clients.storage import StorageArchiveClient, StorageClient, StorageClientFactory
 from py42.exceptions import Py42ArchiveFileNotFoundError
 from py42.response import Py42Response